requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
Django==3.2.11
django-bootstrap4==21.2
django-cors-headers==3.10.1
djangorestframework==3.13.1
numpy==1.22.3
Pillow==9.0.1
pytz==2021.3
soupsieve==2.3.1
sqlparse==0.4.2
